+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `bin` and `accountId` payment fields when the user enters a new credit card.
+- The selected default address is never being sent
 
 ## [0.14.0] - 2021-03-15