+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.14.1] - 2021-03-23
 ### Fixed
 - `bin` and `accountId` payment fields when the user enters a new credit card.
 - The selected default address is never being sent